Moved code to other files
Moved blocks of similar code to their own specific files. For example, all classes are defined within the classes module. All elements of the button class are defined and stored in the elements module.
